backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlmodel>=0.0.14
psycopg[binary]>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
python-multipart>=0.0.6
